requirements.txt
pandas==2.2.2
plotly==5.23.0
scikit-learn==1.5.1
tensorflow==2.17.0
joblib==1.4.2
matplotlib-inline==0.1.7